+++ modules/dxp/apps/ai-hub-nexus/ai-hub-nexus-rest-test/src/testIntegration/java/com/liferay/ai/hub/nexus/rest/resource/v1_0/test/AgentDefinitionResourceTest.java
@@ -98,7 +98,7 @@ public class AgentDefinitionResourceTest
 
 		SiteInitializer siteInitializer =
 			_siteInitializerRegistry.getSiteInitializer(
-				"com.liferay.ai.hub.site.initializer");
+				"com.liferay.ai.hub.nexus.site.initializer");
 
 		siteInitializer.initialize(TestPropsValues.getGroupId());
