+++ modules/dxp/apps/ai-hub-nexus/ai-hub-nexus-rest-test/src/testIntegration/java/com/liferay/ai/hub/nexus/rest/resource/v1_0/test/MessageResourceTest.java
@@ -5,7 +5,7 @@
 
 package com.liferay.ai.hub.nexus.rest.resource.v1_0.test;
 
-import com.liferay.ai.hub.client.security.JWTTokenUtil;
+import com.liferay.ai.hub.cell.security.JWTTokenUtil;
 import com.liferay.ai.hub.nexus.rest.resource.v1_0.test.util.SseEventSourceTestUtil;
 import com.liferay.ai.hub.nexus.rest.resource.v1_0.util.SseUtil;
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
@@ -65,7 +65,7 @@ public class MessageResourceTest extends BaseMessageResourceTestCase {
 
 		SiteInitializer siteInitializer =
 			_siteInitializerRegistry.getSiteInitializer(
-				"com.liferay.ai.hub.site.initializer");
+				"com.liferay.ai.hub.nexus.site.initializer");
 
 		siteInitializer.initialize(TestPropsValues.getGroupId());
 	}
@@ -102,8 +102,8 @@ public class MessageResourceTest extends BaseMessageResourceTestCase {
 			JSONUtil.put(
 				"text", "Expand the following text: " + text
 			).toString(),
-			"ai-hub/v1.0/chats/by-external-reference-code/" + sseEventSinkKey +
-				"/messages",
+			"ai-hub-nexus/v1.0/chats/by-external-reference-code/" +
+				sseEventSinkKey + "/messages",
 			HashMapBuilder.put(
 				"Liferay-AI-Hub-On-Behalf-Of",
 				_generateToken(TestPropsValues.getUserId())
@@ -128,8 +128,8 @@ public class MessageResourceTest extends BaseMessageResourceTestCase {
 			JSONUtil.put(
 				"text", "What was the first message that I sent in this chat?"
 			).toString(),
-			"ai-hub/v1.0/chats/by-external-reference-code/" + sseEventSinkKey +
-				"/messages",
+			"ai-hub-nexus/v1.0/chats/by-external-reference-code/" +
+				sseEventSinkKey + "/messages",
 			HashMapBuilder.put(
 				"Liferay-AI-Hub-On-Behalf-Of",
 				_generateToken(TestPropsValues.getUserId())
