+++ modules/dxp/apps/ai-hub-nexus/ai-hub-nexus-rest-test/src/testIntegration/java/com/liferay/ai/hub/nexus/rest/resource/v1_0/test/util/SseEventSourceTestUtil.java
@@ -69,7 +69,8 @@ public class SseEventSourceTestUtil {
 				).header(
 					"Authorization", authorization
 				).uri(
-					URI.create("http://localhost:8080/o/ai-hub/v1.0/" + uri)
+					URI.create(
+						"http://localhost:8080/o/ai-hub-cell/v1.0/" + uri)
 				).GET(
 				).build(),
 				HttpResponse.BodyHandlers.ofInputStream());
