+++ modules/dxp/apps/ai-hub-nexus/ai-hub-nexus-site-initializer-test/src/testIntegration/java/com/liferay/ai/hub/nexus/site/initializer/internal/test/AIHubSiteInitializerTest.java
@@ -62,7 +62,7 @@ public class AIHubSiteInitializerTest {
 	public void test() throws Exception {
 		SiteInitializer siteInitializer =
 			_siteInitializerRegistry.getSiteInitializer(
-				"com.liferay.ai.hub.site.initializer");
+				"com.liferay.ai.hub.nexus.site.initializer");
 
 		siteInitializer.initialize(TestPropsValues.getGroupId());
