+++ modules/dxp/apps/ai-hub-nexus/ai-hub-nexus-web/src/main/resources/META-INF/resources/edit_agent_definition.jsp
@@ -13,7 +13,7 @@ EditAgentDefinitionDisplayContext editAgentDefinitionDisplayContext = (EditAgent
 
 <div>
 	<react:component
-		module="{AgentDefinitionForm} from ai-hub-web"
+		module="{AgentDefinitionForm} from ai-hub-nexus-web"
 		props="<%= editAgentDefinitionDisplayContext.getReactData() %>"
 	/>
 </div>
